fix(export): Log errors and fail command on errors
Only errors coming from the error object from the `run` method were logged
before. But there might be other kinds of errors coming from webpack and these
need to be handled as well.

Had to add some tests for existing things as well to get past the code coverage
limit that was in the build systems.