Fix select controls to handle groups in reactions (#1516)
* Fix select controls to handle groups in reactions

`options` of `FixtureStateSelectControl` now accepts an array of objects with `group` and `options` in addition to an array of strings. This commit also updates the use of `options` in `useSelect` and `Select` to handle groups, and defines `GroupedOptions` and `isGroupedOptions` in a separate file shared among multiple components. Finally, a new fixture in `Select.fixture.tsx` showcases the new `group`ed options for rendering select controls.

* Address pull request comments

* Stringify grouped options properly

* Use grouped options in example fixture